Criação do task service e modificação nos domains task e project

+++ api/src/main/kotlin/com/api/application/project/exceptions/ProjectExceptions.kt
@@ -1,13 +1,13 @@
 package com.api.application.project.exceptions
 
 sealed class ProjectExceptions(message: String): Exception(message) {
-    abstract val projectID: Int?
+    abstract val projectID: Long?
 }
 
 data class ProjectNotFoundException (
-    override val projectID: Int?
+    override val projectID: Long?
 ): ProjectExceptions("Projeto $projectID não encontrado!")
 
 data class ProjectNotInsertException (
-    override val projectID: Int?
+    override val projectID: Long?
 ): ProjectExceptions("Não foi possivel inserir o projeto de ID: $projectID!")
